refactor(bijou): split dag.ts into dag-layout, dag-edges, dag-render
Extract internal implementation from the 941-line dag.ts into three
focused sibling modules:

- dag-edges.ts: edge routing primitives (Dir, GridState, createGrid,
  markEdge, junctionChar)
- dag-layout.ts: Sugiyama-style layer assignment and column ordering
  (assignLayers, buildLayerArrays, orderColumns)
- dag-render.ts: node box rendering, interactive/pipe/accessible
  renderers (renderInteractiveLayout, renderPipe, renderAccessible)

dag.ts remains the public facade exporting DagNode, DagOptions,
DagNodePosition, DagLayout types and dag(), dagSlice(), dagLayout()
entry points. All imports between new modules use type-only imports
for dag.ts types, preventing runtime circular dependencies.

Zero test modifications. Zero index.ts modifications.
